+++ routes/web.php
@@ -27,25 +27,36 @@ Route::get('/admin/dashboard',[FrontController::class,"admin"])->middleware(['au
 
 Route::get("/admin/users/main",[UserController::class,"index"])->middleware(['auth',"isAdmin"])->name("users.index");
 
-Route::get("/admin/users/create",[UserController::class,"create"])->name("users.create");
+Route::get("/admin/users/create",[UserController::class,"create"])->middleware(['auth',"isAdmin"])->name("users.create");
 
-Route::post("/admin/users/store",[UserController::class,"store"])->name("users.store");
+Route::post("/admin/users/store",[UserController::class,"store"])->middleware(['auth',"isAdmin"])->name("users.store");
 
-Route::get('/admin/users/edit',[UserController::class,"edit"])->name("users.edit");
+Route::get('/admin/users/edit',[UserController::class,"edit"])->middleware(['auth',"isAdmin"])->name("users.edit");
 
-Route::delete("/admin/users/{user}/destroy",[UserController::class,"destroy"])->name("users.destroy");
+Route::delete("/admin/users/{user}/destroy",[UserController::class,"destroy"])->middleware(['auth',"isAdmin"])->name("users.destroy");
 
-Route::put("/admin/users/{user}/update",[UserController::class,"update"])->name("users.update");
+Route::put("/admin/users/{user}/update",[UserController::class,"update"])->middleware(['auth',"isAdmin"])->name("users.update");
 
 Route::put('admin/user/{user}/edit', [UserController::class, 'updateMembre'])->name('membre.update');
 
+//Avatar
 Route::resource("admin/avatar",AvatarController::class)->middleware(["auth","isAdmin"]);
 
+// Cateories
 Route::resource('/admin/categorie', CategorieController::class)->middleware(["auth","isAdmin"]);
 
+// Images
 Route::resource('/admin/image', ImageController::class)->middleware(["auth","isAdmin"]);
 
-Route::get('/admin/blog', [ArticleController::class, 'blog'])->name('blog.index');
+// Article Blog
+Route::get('/admin/blog', [ArticleController::class, 'blog'])->middleware(["auth","isAdmin"])->name('blog.index');
 
+//article
 Route::resource('admin/article', ArticleController::class);
+
+// Gallerie
+Route::get('/admin/gallerie', [ImageController::class, 'gallerie'])->name('gallerie.index');
+
+Route::get('/admin/{image}/download', [ImageController::class, 'download'])->name('gallerie.download');
+
 require __DIR__.'/auth.php';
